HtmlArea in an item-set is not being updated #7716

+++ modules/lib/src/main/resources/assets/js/app/inputtype/text/HtmlArea.ts
@@ -35,6 +35,7 @@ import {HtmlEditor} from '../ui/text/HtmlEditor';
 import {HtmlEditorParams} from '../ui/text/HtmlEditorParams';
 import {StylesRequest} from '../ui/text/styles/StylesRequest';
 import {HtmlAreaResizeEvent} from './HtmlAreaResizeEvent';
+import {DefaultErrorHandler} from '@enonic/lib-admin-ui/DefaultErrorHandler';
 
 export class HtmlArea
     extends BaseInputTypeNotManagingAdd {
@@ -116,6 +117,14 @@ export class HtmlArea
 
             this.onOccurrenceRemoved(() => this.resetInputHeight());
         });
+
+        this.onOccurrenceValueChanged((occurrence: TextAreaWrapper, value: Value) => {
+            const editor = this.editors.find((e: HtmlAreaOccurrenceInfo) => e.textAreaWrapper === occurrence);
+
+            if (editor) {
+                editor.savedValue = value.getString();
+            }
+        });
     }
 
     getValueType(): ValueType {
@@ -139,12 +148,18 @@ export class HtmlArea
         textAreaEl.addClass(clazz);
 
         const textAreaWrapper = new TextAreaWrapper('text-area-wrapper');
+        const editor = {
+            id: editorId,
+            textAreaWrapper,
+            textAreaEl,
+            savedValue: property.hasNonNullValue() ? property.getString() : '',
+            hasStickyToolbar: false
+        };
+        this.editors.push(editor);
 
         textAreaEl.onRendered(() => {
             this.authRequest.then(() => {
-                this.initEditor(editorId, property, textAreaWrapper).then(() => {
-                    this.editors.push({id: editorId, textAreaWrapper, textAreaEl, property, hasStickyToolbar: false});
-                });
+                this.initEditor(editorId, editor.savedValue, textAreaWrapper).catch(DefaultErrorHandler.handle);
             });
         });
 
@@ -166,11 +181,13 @@ export class HtmlArea
 
     protected updateFormInputElValue(occurrence: FormInputEl, property: Property) {
         const editor: HtmlAreaOccurrenceInfo = this.getEditorInfo(occurrence.getId());
+        const newValue = property.hasNonNullValue() ? property.getString() : '';
+
         if (editor) {
-            editor.property = property;
+            editor.savedValue = newValue;
         }
 
-        this.setEditorContent(occurrence as TextArea, property);
+        this.setEditorContent(occurrence as TextArea, newValue);
     }
 
     resetInputOccurrenceElement(occurrence: Element) {
@@ -211,7 +228,7 @@ export class HtmlArea
         iframe?.setAttribute('tabindex', enable? '0' : '-1');
     }
 
-    private initEditor(id: string, property: Property, textAreaWrapper: Element): Q.Promise<HtmlEditor> {
+    private initEditor(id: string, value: string, textAreaWrapper: Element): Q.Promise<HtmlEditor> {
         const focusHandler = (e) => {
             this.resetInputHeight();
             this.notifyFocused(e);
@@ -294,7 +311,7 @@ export class HtmlArea
         };
 
         const editorReadyHandler = (eventInfo: CKEDITOR.eventInfo) => {
-            this.setEditorContent(textAreaWrapper.findChildById(id) as TextArea, property);
+            this.setEditorContent(textAreaWrapper.findChildById(id) as TextArea, value);
             const editor = this.editors.find((editor: HtmlAreaOccurrenceInfo) => editor.id === id);
 
             if (editor && !this.enabled) {
@@ -469,13 +486,11 @@ export class HtmlArea
         }
     }
 
-    private setEditorContent(textArea: TextArea, property: Property): void {
+    private setEditorContent(textArea: TextArea, value: string): void {
         const editorId: string = textArea.getId();
-        const content: string = property.hasNonNullValue() ?
-                                HTMLAreaHelper.convertRenderSrcToPreviewSrc(property.getString(), this.content?.getId(),
-                                    this.context.project) : '';
 
         if (HtmlEditor.exists(editorId)) {
+            const content: string = HTMLAreaHelper.convertRenderSrcToPreviewSrc(value, this.content?.getId(), this.context.project);
             const currentData: string = HtmlEditor.getData(editorId);
             // invoke setData only if data changed
             if (content !== currentData) {
@@ -496,6 +511,7 @@ export class HtmlArea
 
     private handleEditorValueChanged(id: string, occurrence: Element) {
         const value: string = HtmlEditor.getData(id);
+
         const textAreaEl = occurrence.findChildById(id) as TextArea;
         if (textAreaEl && value !== textAreaEl.getValue()) {
             textAreaEl.setValue(value, false, true);
@@ -527,8 +543,10 @@ export class HtmlArea
         this.editors.forEach((editor) => {
             const editorId = editor.id;
 
-            this.destroyEditor(editorId);
-            this.reInitEditor(editorId);
+            if (HtmlEditor.exists(editorId)) {
+                this.destroyEditor(editorId);
+                this.reInitEditor(editorId);
+            }
         });
     }
 
@@ -583,7 +601,7 @@ export class HtmlArea
     private reInitEditor(id: string): Q.Promise<HtmlEditor> {
         const savedEditor: HtmlAreaOccurrenceInfo = this.getEditorInfo(id);
 
-        return this.initEditor(id, savedEditor.property, savedEditor.textAreaWrapper);
+        return this.initEditor(id, savedEditor.savedValue, savedEditor.textAreaWrapper);
     }
 
     private getEditorInfo(id: string): HtmlAreaOccurrenceInfo {
@@ -615,7 +633,7 @@ export interface HtmlAreaOccurrenceInfo {
     id: string;
     textAreaWrapper: Element;
     textAreaEl: TextArea;
-    property: Property;
+    savedValue: string;
     hasStickyToolbar: boolean;
 }
